% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_calculate_coeff_r.R
\name{calculate_coeff_r}
\alias{calculate_coeff_r}
\title{Calculate the coefficient of reliability (R).}
\usage{
calculate_coeff_r(total_tem, sd)
}
\arguments{
\item{total_tem}{A numeric value for total TEM}

\item{sd}{A numeric vector for standard deviation of measurements}
}
\value{
A numeric vector of coefficient of reliability values per observer
}
\description{
Calculate the coefficient of reliability (R).
}
\examples{
#


}
