% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_calculate_team_tem.R
\name{calculate_team_tem}
\alias{calculate_team_tem}
\title{calculate_team_tem}
\usage{
calculate_team_tem(n, k, m)
}
\arguments{
\item{n}{A numeric value specifying number of subjects}

\item{k}{A numeric value specifying number of observers}

\item{m}{A data frame of of measurements with number of columns equal to
\code{k} and number of rows equal to \code{n}}
}
\value{
A numeric value of the technical error of measurement (TEM). Unit
    of TEM is the same as the unit of the measurements compared.
}
\description{
Function to calculate technical error of measurement (TEM) when there are
more than two observers involved.
}
\examples{
# calculate team TEM from stature dataset
calculate_team_tem(n = 10, k = 4, m = stature[ , c("m1", "m2", "m3", "m4")])


}
