% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_team_tem.R
\name{calculate_team_tem_}
\alias{calculate_team_tem_}
\title{calculate_team_tem}
\usage{
calculate_team_tem_(n, k, m)
}
\arguments{
\item{n}{A numeric value specifying number of subjects}

\item{k}{A numeric value specifying number of observers}

\item{m}{A data frame of of measurements with number of columns equal to
\code{k} and number of rows equal to \code{n}}
}
\value{
A numeric value of the technical error of measurement (TEM). Unit
    of TEM is the same as the unit of the measurements compared.
}
\description{
Function to calculate technical error of measurement (TEM) when there are
more than two observers are involved.
}
\examples{
#

}
