% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculate_tem_cohort.R
\name{calculate_tem_cohort}
\alias{calculate_tem_cohort}
\title{Calculate intra-observer TEM for each observer}
\usage{
calculate_tem_cohort(df, m1, m2, index = NULL, n)
}
\arguments{
\item{df}{A data.frame containing data on multiple measurements made by
multiple observers on multiple subjects}

\item{m1}{A character value for variable in `df` containing initial round
measurements made by multiple observers}

\item{m2}{A character value for variable in `df` containing subsequent round
measurements made by multiple observers}

\item{index}{A character value or vector specifying the variable name/s in
`df` containing the factors to be used to group the results into. If
multiple observers, at least two variable  names should be
provided as observer identifiers. Any other variable in `df` that should
be used as grouping factor should be included. This is *optional* and is
set to NULL by default. This would indicate that the summary metrics
will be made for the values in `measures` without any grouping}

\item{n}{Number of subjects measured}
}
\value{
A data frame of calculated TEM per observer and per
    measurement type.
}
\description{
Function to calculate *intra-observer TEM* for each observer using input
dataset containing multiple types of multiple measurements from multiple
observers.
}
\examples{
# Get the TEM per observer for MUAC measurements in the smartStd dataset
calculate_tem_cohort(df = smartStd,
                     m1 = "muac1", m2 = "muac2",
                     index = "observer", n = 10)

}
