% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_estimate_bias.R
\name{estimate_bias}
\alias{estimate_bias}
\title{estimate_bias}
\usage{
estimate_bias(msur, msup, mall)
}
\arguments{
\item{msur}{A numeric value or vector of mean measurements from observers}

\item{msup}{A numeric value or vector of mean measurements from supervisor}

\item{mall}{A numeric value or vector of the median of  measurements from
observers}
}
\value{
A numeric value or vector signifying bias
}
\description{
Calculate the bias of a measurement against a gold standard. Two gold standards
are used in this function: 1) measurements made by a supervisor or a known
expert; and, 2) median of all measurements made by the observers.
}
\examples{
#

}
