% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liberiaStdData}
\alias{liberiaStdData}
\title{Dataset from a standardisation exercise done in Liberia in preparation for
a coverage survey.}
\format{
A data frame with 7 columns and 744 rows:
\describe{
\item{\code{eid}}{Enumerator ID}
\item{\code{cid}}{Child ID}
\item{\code{round}}{Standardisation test round}
\item{\code{weight}}{Weight measurement}
\item{\code{height}}{Height measurement}
\item{\code{muac}}{Mid-upper arm circumference measurement}
\item{\code{oedema}}{Bilateral pitting oedema}
}
}
\usage{
liberiaStdData
}
\description{
Dataset from a standardisation exercise done in Liberia in preparation for
a coverage survey.
}
\keyword{datasets}
