% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_runApp.R
\name{run_anthrocheckr}
\alias{run_anthrocheckr}
\title{Run the anthrocheckr Shiny app}
\usage{
run_anthrocheckr()
}
\description{
Run the anthrocheckr Shiny app
}
\examples{
if (interactive()) run_anthrocheckr()

}
