% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smartWide}
\alias{smartWide}
\title{Standardisation test data - sample 3}
\format{
A data frame with 67 columns and 10 rows
}
\source{
Standardized Monitoring and Assessment of Relief and Transitions
    (SMART) capacity building toolbox found at
    <http://smartmethodology.org/survey-planning-tools/smart-capacity-building-toolbox/>.
}
\usage{
smartWide
}
\description{
Example dataset from the Standardized Monitoring and Assessment of Relief
and Transitions (SMART) capacity building toolbox found at
<http://smartmethodology.org/survey-planning-tools/smart-capacity-building-toolbox/>.
This dataset is the same as the \code{smartStd} dataset but has been kept in
the original wide format.
}
\examples{
smartWide

}
\keyword{datasets}
