% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_summary_measure.R
\name{summary_measure}
\alias{summary_measure}
\title{Calculate the summary measures of observations}
\usage{
summary_measure(df, measures, index = NULL)
}
\arguments{
\item{df}{A data.frame containing observations/measurements made by
a single observer or multiple observers across multiple subjects}

\item{measures}{A character value specifying the variable name in `df`
containing vector of measurements made by an observer or multiple
observers on multiple subjects.}

\item{index}{A character value or vector specifying the variable name/s in
`df` containing the list of factors to be used to group the summary results
into. If multiple observers, at least two variable  names should be
provided as observer identifiers. Any other variable in `df` that should
be used as grouping factor should be included. This is *optional* and is
set to NULL by default. This would indicate that the summary metrics
will be made for the values in `measures` without any grouping.}
}
\value{
A data frame containing calculated summary metrics of the
  specified `measures` grouped by specified `index`.
}
\description{
Get the mean, standard deviation and maximum value of the observations/
measurements made by an observer across multiple subjects
}
\examples{
# Apply summary_measure on smartStd dataset that has a row for each subject
# measurement and multiple columns of different measurement types
# (i.e, height, weight, muac) taken at different times (repeat measurements).
# The grouping variable is the observer column. This can be used to get the
# summary measure for one measurement type only but can be used to summarise
# repeat measurements
# Get mean of weight measure taken by each observer at 2 separate occasions
  x <- smartStd[ , c("observer", "weight1")]
  y <- smartStd[ , c("observer", "weight2")]
  names(x) <- names(y) <- c("observer", "weight")
  temp <- data.frame(rbind(x, y))
  mean_measure <- summary_measure(df = temp,
                                  measures = "weight",
                                  index = "observer")
  mean_measure

# Apply summary_measure on smartStdLong dataset that has a row for each
# subject for each measure type and for each repeat measurement. Get mean,
# sd and max of height, weight and MUAC in one specification
summary_measure(df = smartStdLong,
                measures = "measure_value",
                index = c("observer", "measure_type"))

}
