% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_anthrocheckr.R
\docType{package}
\name{anthrocheckr}
\alias{anthrocheckr-package}
\alias{anthrocheckr}
\title{An Implementation of Anthropometric Measurement Standardisation Tests}
\description{
Ensuring the precision and accuracy of measurements is critical when
collecting anthropometric data. Anthropometrists are usually tested for
precision and accuracy of measurement through standardisation tests performed
prior to anthropometric data collection. This package provides functions to
calculate inter- and intra-observer technical error of measurement (TEM) to
assess precision of measurements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nutriverse.io/anthrocheckr/}
  \item \url{https://github.com/nutriverse/anthrocheckr}
  \item Report bugs at \url{https://github.com/nutriverse/anthrocheckr/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
