% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tem_cohort.R
\name{calculate_tem_cohort}
\alias{calculate_tem_cohort}
\title{calculate_tem_cohort}
\usage{
calculate_tem_cohort(x, d, n)
}
\arguments{
\item{x}{A data frame in long format containing data on multiple measurements
made by multiple observers on multiple subjects}
}
\description{
Function to calculate intra-observer TEM for each observer using input
dataset containing multiple types of multiple measurements from multiple
observers.
}
