% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculate_tem_cohort.R
\name{calculate_tem_cohort}
\alias{calculate_tem_cohort}
\title{Calculate intra-observer TEM for each observer}
\usage{
calculate_tem_cohort(m1, m2, index, n)
}
\arguments{
\item{m1}{A numeric vector of first round measurements made by multiple
observers}

\item{m2}{A numeric vector of second round measurements made by multiple
observers}

\item{index}{A list of grouping factors}

\item{n}{Number of subjects}
}
\value{
A vector or data frame of calculated TEM per observer and per
    measurement type.
}
\description{
Function to calculate intra-observer TEM for each observer using input
dataset containing multiple types of multiple measurements from multiple
observers.
}
\examples{
# Get the TEM per observer for MUAC measurements in the smartStd dataset
calculate_tem_cohort(m1 = smartStd$muac1, m2 = smartStd$muac2,
                     index = smartStd$observer, n = 10)

# Get the TEM per observer for all measurement types (height, weight, MUAC)
# in the smartStdLong dataset
x <- smartStdLong[smartStdLong$measure_round == 1, ]
y <- smartStdLong[smartStdLong$measure_round == 2, ]

calculate_tem_cohort(m1 = x$measure_value, m2 = x$measure_value,
  index = x[ , c("observer", "measure_type")],
  n = 10)

}
