% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_estimate_bias.R
\name{estimate_bias}
\alias{estimate_bias}
\title{Calculate bias of a measurement}
\usage{
estimate_bias(msur, msup, mall)
}
\arguments{
\item{msur}{A numeric value or vector of mean measurements from observers}

\item{msup}{A numeric value or vector of mean measurements from supervisor}

\item{mall}{A numeric value or vector of the median of measurements from
observers}
}
\value{
A numeric value or vector signifying bias
}
\description{
Calculate the bias of a measurement against a gold standard. Two gold
standards are used in this function: 1) measurements made by a supervisor or
a known expert; and, 2) median of all measurements made by the observers.
}
\examples{
x <- summary_measure(df = smartStdLong,
                     measures = "measure_value",
                     index = c("observer", "measure_type"))
y <- summary_measure(df = smartStdLong,
                     measures = "measure_value",
                     index = "measure_type")
estimate_bias(msur = x$mean[x$observer != 0 & x$measure_type == "height"],
              msup = x$mean[x$observer == 0 & x$measure_type == "height"],
              mall = y$mean[y$measure_type == "height"])

}
