% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liberiaStdData}
\alias{liberiaStdData}
\title{Dataset from a standardisation exercise done in Liberia in preparation for
a coverage survey.}
\format{
A data frame with 7 columns and 744 rows:

| **Variable** | **Description** |
| :--- | :--- |
| *eid* | Enumerator ID |
| *cid* | Child ID |
| *round* | Standardisation test round |
| *weight* | Weight measurement |
| *height* | Height measurement |
| *muac* | Mid-upper arm circumference measurement |
| *oedema* | Bilateral pitting oedema |
}
\usage{
liberiaStdData
}
\description{
Dataset from a standardisation exercise done in Liberia in preparation for
a coverage survey.
}
\examples{
liberiaStdData

}
\keyword{datasets}
